PokemonBackend/requirements.txt
flask
pymongo
python-dotenv
flasgger